+++ tests/src/Opulence/Validation/Rules/EqualsRuleTest.php
@@ -8,6 +8,8 @@
  */
 namespace Opulence\Validation\Rules;
 
+use InvalidArgumentException;
+
 /**
  * Tests the equals rule
  */
@@ -18,8 +20,28 @@ class EqualsRuleTest extends \PHPUnit_Framework_TestCase
      */
     public function testEqualValuesPass()
     {
-        $rule = new EqualsRule("foo");
+        $rule = new EqualsRule();
+        $rule->setArgs(["foo"]);
         $this->assertTrue($rule->passes("foo"));
+    }
+
+    /**
+     * Tests getting the slug
+     */
+    public function testGettingSlug()
+    {
+        $rule = new EqualsRule();
+        $this->assertEquals("equals", $rule->getSlug());
+    }
+
+    /**
+     * Tests passing an empty arg array
+     */
+    public function testPassingEmptyArgArray()
+    {
+        $this->setExpectedException(InvalidArgumentException::class);
+        $rule = new EqualsRule();
+        $rule->setArgs([]);
     }
 
     /**
@@ -27,7 +49,8 @@ class EqualsRuleTest extends \PHPUnit_Framework_TestCase
      */
     public function testUnequalValuesFail()
     {
-        $rule = new EqualsRule("foo");
+        $rule = new EqualsRule();
+        $rule->setArgs(["foo"]);
         $this->assertFalse($rule->passes("bar"));
     }
 }